edited sysagnt powershell code
We worked on and experimented with formatting the output of this code.
We decided to have it create a list of attributes of the Windows OS and
create a list with attributes delimited by commas, and a newline at the
end of each string, rather than have a newline after each process and
application like it was before. Our current idea is to pipe it through
PHP code which will then organize and format (and output) this data in a
concise and logical way.